refactor: use a single index page regardless user type
